requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
anthropic==0.40.0
pydantic==2.9.2
python-dotenv==1.0.1

# Voice agent
livekit-agents>=1.0
livekit-plugins-openai
livekit-plugins-silero
livekit-api
